commit before norm

+++ libft/ft_strnew.c
@@ -6,7 +6,7 @@
 /*   By: yismail <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/01 13:26:26 by yismail           #+#    #+#             */
-/*   Updated: 2015/12/01 13:41:05 by yismail          ###   ########.fr       */
+/*   Updated: 2015/12/02 15:18:54 by yismail          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strnew(size_t size)
 {
 	char *str;
-	int i;
+	size_t i;
 
 	i = 0;
 	str = (char*)malloc(sizeof(char) * size);
